move vaildator model and migrations to core
