+++ dataactcore/alembic.ini
@@ -32,7 +32,7 @@ script_location = migrations
 # list of database names to target in the migration. set to the complete
 # list of broker backend db names. individual db creation scripts
 # overwrite this setting as necessary
-databases = error_data, job_tracker, user_manager
+databases = error_data, job_tracker, user_manager, validation
 
 # [error_data]
 # sqlalchemy.url = ''
@@ -41,6 +41,9 @@ databases = error_data, job_tracker, user_manager
 # sqlalchemy.url = ''
 
 # [user_manager]
+# sqlalchemy.url = ''
+
+# [validation]
 # sqlalchemy.url = ''
 
 # Logging configuration
