--- dataactvalidator/validator_configuration_example.json
@@ -1,1 +1,0 @@
-{"rest_trace": false, "port": "80", "server_debug": false}
